+++ src/Home.js
@@ -1,4 +1,4 @@
-import { Header } from '../Components/Header';
+import { Header } from './Components/Header';
 import React from "react";
